Format macro def with repeat
This commit lets rustfmt to format macro def with repeat in it.

The strategy is to  replace `$(..)*` or `$(..)+` with `({..})` and proceed
formatting. This works well when repeated content is or statements.
The downside is that we fail to format when the repeated content ends with
a trailing comma.

This commit is *VERY* messy...
I am plannig to refactor this and add more tests.