Don't reorder modules by default
We can flip this, although this is stable, because disabling
reordering won't change existing order.

Why do we disabling this?

The problem with alphabetical ordering is that is as good as any other
arbitrary ordering. It does not have intrinsic benefits, and it is not
really about enforcing consistency: you only repeat a particular
module once, so any order is consistent.

The default zero-effort ordering of "modules are ordered according to
git history" is better than alphabetical, because it makes git blame
more consistent and gives a hint about what modules are older.

The significantly superior ordering, which probably should be a part
of style guide, is "suggested reading order". We obviously can't
enforce it via rustfmt, so let's just not reorder imports at all.