Exclude where kw from comment following fn return type (#4175)
rustfmt tries to preserve the comment between a fn return type and the
start of the fn body if there are no where clauses following the return
type. However, even if there are no where clauses present, the "where"
keyword may be. To elide the "where" keyword in this context, just get
the comment snippet starting after the where clause span, which always
includes the "where" keyword if present.

Closes #4001