Fix module resolution in inner modules with paths
Path attributes with relative paths were previously not treated as
relative to the containing file but as relative to the current working
directory of rustfmt.

The SubModKind enum had a dedicated variant for inner modules with
paths. However, the ordinary variant for inner modules also handles the
presence of path attributes and treats them correctly. Therefore the fix
is to simply remove the dedicated enum variant.

Fixes #3901
Fixes #4076