Drop trailing semicolons on item statements
Per the Rust grammar, trailing semicolons on an item statement may be
optional (i.e. if the statement is a block). Keeping the semicolon in
such cases is unnecessary and may be confusing (does the semicolon
terminate the last statement, or is it an empty statement?), so just
drop it.

Closes #4222