Pick up comments between trait where clause and open block
Previously rustfmt only picked up comments between a trait "signature"
and its block when there was no `where` clause on the trait. Amending
this to work for comments following a `where` clause and preceding the
trait block is straightforward; we simply get rid of the branch and
adjust the span to search for a comment in accordingly.

Closes #4289