Permit ranges on multiple lines in closures, without an enclosing block
The problem here was that if an expression in a closure fails to be
formatted without an enclosing block, as in the case of a multiline
range prior to this commit, we try to put the expression in the block.
But then we try to check what expressions should be "vetoed" from being
in a block, which includes ranges. The non-breaking fix is to let
multiline ranges be formatted without a block.

On the big picture, though, it may be the case that there is no need for
a "veto" check when formatting expressions in a closure block, or that
there is a better way to structure the work being done here. I added a
todo for this, since it's relatively low priority (IMO).

Closes #4308