Fix idempotency issue when wrapping the last match arm in a block
The issues was related to the interaction between two differnt
configuration options when applied to the last match arm.

* trailing_comma = Never
* match_block_trailing_comma = true

Previously, rustfmt only consider the ``match_block_trailing_comma``
configuration value when wrapping match arms. This lead to rustfmt
adding a trailing comma to all newly wrapped match arms.

Now, rustfmt also considers the ``trailing_comma`` configuration value
when deciding whether or not to add a trailing comma when wrapping the
last match arm in a block.