`rustfmt`: Use `edition` form `Cargo.toml` rather than always `2015` (DO NOT MERGE)
This patch shows an example solution to #4074, where `rustfmt` will fetch the `edition` from the source of truth: `Cargo.toml`. This is meant as a point of discussion rather than a serious fix.

For example, this allows `rustfmt` to be run against files in the `rustfmt` repository itself and automatically detect the correct edition of `2018` as configured in Cargo.toml, rather than using the incorrect default of `2015` like it does today.

From discussion in #4074, it sounds like the current preference is to not have `rustfmt` detect the `edition` automatically and to instead have a different command for end users to format individual files, but in the meantime this provides a reference for one way that it could be done.

Note that if this were a serious PR I would also have written tests for the change, including adding some sorely needed coverage of `resolve_config_files` that is currently missing, but again this is just meant to be an example to show how the lookup logic could work.

As of this patch, the `edition` selection order is in the following priority:
1. `--edition` provided as an argument
2. `edition` in "nearest" `rustfmt.toml` to the file
3. `package.edition` in "nearest" `Cargo.toml` to the file (NEW!)
4. `edition` in system `rustfmt.toml`s (IMO this is wrong - causes problems when different repos have different editions)
5. `2015`