Tie inner attributes to the correct `#[cfg(..)]` mod
It's possible that multiple modules could have the same name. E.g.:

    #[cfg(linux)]
    mod foo;
    #[cfg(not(linux))]
    mod foo;

Previously, we would just assign attributes to the first module we found
with the same name as the one we were trying to associate inner
attributes with.This lead to a situation where we assigned all inner sub
module attributes to the first module we found.

Now we also ensure the spans for each item match so that we know we're
associating the inner attributes with the correct mod item.