Use the same shape to rewrite import when reordering them or not
Fixes 4794

The code paths for formatting imports diverge based on the value of
`reorder_imports`. Each code path used a slightly different shape when
rewriting the imports which lead to issues when `reorder_imports = false`
and `imports_indent = Visual`.

Now the code path that rewrites imports without reordering uses the same
shape as the path that also reorders imports.