Add `fn_call_layout` configuration option
Closes 5218

The `fn_call_layout` was added to give users more control over how
function calls are formatted.

This change will only impact function calls, and will not have any
affect on method calls.

The values are identical to those for `fn_args_layout`, which  is a
stable option that controls how function declarations are laid out.