group_imports: add `ExternalCrate`
This is similar to the existing `StdExternalCrate`, but doesn't give
`std`, `core`, and `alloc` their own group. Conceptually, `std` and
friends are just external libraries (though they have some magical
abilities thanks to lang items and `rustc_attrs`), so this is the style
I naturally gravitated towards. I'm not sure how common this pattern is
throughout the ecosystem, but I do hope this will be useful to someone
other than me.