Wrap cast expr in parens when the ty ends with empty generic args
Fixes 4621

Previously rustfmt would remove empty generic args from cast
expressions e.g. `x as i32<>` => `x as i32`. This lead to code that
could not compile when the cast occurred in the context of a binary
expression where the operand was a less than (`<`) or left shift (`<<`).

The advice from the parse error emitted by the compiler was to wrap the
cast in parentheses, which is now the behavior rustfmt employs.

* `x as i32<> < y` => `(x as i32) < y`
* `x as i32<> << y` => `(x as i32) << y`