Do no memoize `ast::ExprKind::MethodCall`
Fixes 5399

Memoizing method calls lead to cases where rustfmt's stability guarantees
were violated.

Excluding method calls is a small performance hit that we'll gladly
pay to make sure we maintain our stability guarantees.

As of now it's unclear if there will be other `ast::ExprKind` nodes that
will need to be excluded for a similar reason.