Add new config option `OverflowDensity`
Through testing it was found that `Tall` as the default value for
`fn_call_layout` was not backwards compatable and would lead to
breaking formatting changes.

Instead of version gating the new behavior, which would also lead to
unexpected formatting changes for users who already set `version=Two`,
it was decided that a new option should be add to encapsulate rustfmt's
previous function argument behavior.

In the context of `fn_call_layout` this new option has all the same
variants as the `Density` option used for `fn_params_layout`.