Don't duplicate line numbers when annotating internal errors
Closes 5463

The `annotate_snippets` crate adds line number and column info to the
output, so we don't need to append the line number to the `origin`.

Previously, internal errors would be output with the line number twice.
See how 478 is duplicated in the example below.

```
error[internal]: line formatted, but exceeded maximum width
   --> /path/to/file.rs:478:478:101
```

Now the line number is not duplicated.

```
error[internal]: line formatted, but exceeded maximum width
   --> /path/to/file.rs:478:101
```