Add missing line to use_field_init_shorthand "true" example
The last declaration (`let b = ...`) is where all the difference between the
"false" and "true" cases lies, but it was omitted from the "true" example for
unknown reasons.

Assuming it was by accident, this patch adds the last line from the "false"
example to the "true" example, formatted as it would have been with
`use_field_init_shorthand = true`.