Set `ctx.inside_macro` based on `FmtVisitor.parent_context`
Fixes 5526

When creating new contexts with `FmtVisitor::get_context` we set the
value of `ctx.inside_macro` based on the `parent_context` if it exists.
This helps us remember that we're inside of a macro when creating new
`RewriteContext` from recursively created `FmtVisitors`.

A similar issue to 5526 was reported in 3139, and remembering that we're
inside of a macro in this way helps us go down a code path that resolved
3139 and was implemented in PR 3142.