implement formatting code blocks in markdown files
The `rewrite_markdown` function will take markdown as input and apply
rustfmt on rust code blocks. Because of how `pulldown-cmark-to-cmark`
is implemented there will be some light rewriting of the markdown file
outside of code block reformatting. This mostly relates to updating
newlines after markdown elements like headers and lists, and possibly
updating the indentation of indented code blocks.

In the future we may even consider implementing doc comment rewriting in
terms of `rewrite_markdown`.