Add StyleEditionDefault trait, StyleEdition enum, and #[style_edition]
The StyleEditionDefault trait implement the foundation for the
`style_edition` mechanism. Eventually all configs will be required
to implement this trait so rustfmt can dynamically determine what
default value to assign for a given style edition.

A proc macro is also added so types can conveniently implement the
new trait.