fix(items): Fix handling of rustdoc and macro attributes in enum
This fix was made thanks to the hint at [1].
This was reported in issue #5662 [2].

Previously, a enum item containing an attribute (rustdoc or macro) would
be considered multi-line, thus forcing the formatting strategy of all
the items in the enum to be Vertical (i.e. multi-line formatting).

When determining the formatting strategy for enum items, we should
ignore the attributes. This is what we do in the `is_multi_line_variant`
function. Or else, simply adding a rustdoc comment or a macro attribute
would cause the formatting of the whole enum to change, which is not a
desirable behavior.

We will be adding tests in the following commits.

- [1] https://github.com/rust-lang/rustfmt/issues/5662#issuecomment-1386231624
- [2] https://github.com/rust-lang/rustfmt/issues/5662