Suppress warning if max_width reduced below other settings in macro
The warning makes sense at the top level because it indicates a user
misconfiguration, but the warning can also fire when a macro reduces
max_width within its scope, which is confusing.  We can plumb a flag
through saying whether to warn or not and have the macro-related
twiddling disable the warning, which will still be on at the top level.